remove the timestamp estimator completely
The DVL is essentially a latency-free sensor, and on our system at
least, has very little jitter on top.

However, it's not fully periodic when the bottom lock gets lost. This
(rightly) confuses the timestamp estimator, which drifts and causes
havoc even when the samples are received again.

Given that it seems that the estimator itself has very little benefit
w.r.t the quality of the timestamping, disable it.